requirements.txt
fastapi
uvicorn
playwright
python-dotenv
openrouter
